requirements.txt
aioredis==1.3.1
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.2.6
django-cors-headers==3.1.1
django-rest-framework==0.1.0
djangorestframework==3.10.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
hiredis==1.0.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-http-client==3.2.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
sendgrid==6.1.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
uritemplate==3.0.1
urllib3==1.25.8
zope.interface==4.7.2
